adding tutorials and examples
